Replace framework line with unminified version
Mozilla wants unminified code for review, so we replace at build time.
The version inserted by Scaffold no longer matters -- all versions of
the framework are now replaced by the version from the submodule, so the
submodule will need to be updated if the framework changed.

If older framework versions aren't compatible with the latest version,
a few older translators might need to be updated.