Remove generate-schema script
This script was used to generate the schema file from the types and
fields in the database, but now the schema file itself is the
authoritative record, and further changes to types and fields should be
made directly. Localized strings can be updated by the update-schema
script (adapted from this one) in the 'scripts' directory of the
zotero-schema submodule in zotero/zotero.